requirements.txt
# Runtime（Web + Spider + 离线处理）
Flask==3.1.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.44
PyMySQL==1.1.2

# HTTP / Spider
requests==2.32.5
urllib3==2.2.3

# 图片处理（头像上传/缩略）
Pillow==11.0.0

# 本地开发：用 .env 管理敏感配置（可选但推荐）
python-dotenv==1.0.1